requirements.txt
python-telegram-bot
openai>=1.0.0
python-dotenv